added navigation between pages, please please please please please deploy
